gateway凍結/requirements.txt
# =========================================================
# ASSIST_KEY: このファイルは【gateway/requirements.txt】に位置するユニットです
# =========================================================
#
# 【概要】
#   Gateway サービスで必要な Python 依存パッケージ一覧
#
# 【ルール遵守】
#   1) バージョンは “最低保証 (>=)” で固定、フリーズは CI で行う
#   2) 不要ライブラリを入れない（Slim イメージ維持）
# ---------------------------------------------------------

fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
jinja2>=3.1
PyJWT>=2.8
pydantic>=2.7
typing_extensions>=4.11      # FIXME: Python 3.11 で不要になったら削除
